+++ .storybook/config.js
@@ -1,10 +1,22 @@
+import React from "react";
+
 import { configure, addDecorator } from "@storybook/react";
 import { setDefaults } from "@storybook/addon-info";
 import { setOptions } from "@storybook/addon-options";
 import { withKnobs } from "@storybook/addon-knobs/react";
-import { withThemesProvider } from 'storybook-addon-styled-component-theme';
+import { withThemesProvider } from "storybook-addon-styled-component-theme";
+// import { checkA11y } from "@storybook/addon-a11y";
+import backgrounds from "@storybook/addon-backgrounds";
 
-import { keen, keenDark } from 'style/styleVariables';
+import { keen, keenDark } from "style/styleVariables";
+import PropTypesTable from "./PropTypesTable";
+import { StoryContainer } from "ComponentLibrary/StoryComponents/StoryContainer";
+
+import {
+  COLOR_BACKGROUND_A,
+  COLOR_BACKGROUND_B,
+  COLOR_BACKGROUND_C
+} from "./../src/style/styleVariables";
 
 const req = require.context("../src", true, /\.stories\.js$/);
 
@@ -12,7 +24,15 @@ function loadStories() {
   req.keys().forEach(filename => req(filename));
 }
 // add withKnobs globally so we don't need to import it into individual stories
+// addDecorator(checkA11y);
 addDecorator(withKnobs);
+addDecorator(
+  backgrounds([
+    { name: "COLOR_BACKGROUND_A", value: COLOR_BACKGROUND_A, default: true },
+    { name: "COLOR_BACKGROUND_B", value: COLOR_BACKGROUND_B },
+    { name: "COLOR_BACKGROUND_C", value: COLOR_BACKGROUND_C }
+  ])
+);
 // addon-info
 setDefaults({
   header: true, // Toggles display of header with component name and description
@@ -22,7 +42,15 @@ setDefaults({
     /* Components used in story */
   ], // displays Prop Tables with this components
   propTablesExclude: [], // Exclude Components from being shoun in Prop Tables section
-  styles: {}, // Overrides styles of addon
+  styles: {
+    infoBody: {
+      boxShadow: "inset 0 0 0 rgba(0,0,0,0.08)",
+      border: "0",
+      borderRadius: "8px",
+      backgroundColor: "rgba(200,200,200,.2)"
+    }
+  }, // Overrides styles of addon
+  TableComponent: PropTypesTable,
   components: {}, // Overrides components used to display markdown. Warning! This option's name will be likely deprecated in favor to "components" with the same API in 3.3 release. Follow this PR #1501 for details
   maxPropsIntoLine: 1, // Max props to display per line in source code
   maxPropObjectKeys: 10,
@@ -31,6 +59,7 @@ setDefaults({
 });
 
 addDecorator(withThemesProvider([keen, keenDark]));
+addDecorator(story => <StoryContainer>{story()}</StoryContainer>);
 
 // Options:
 setOptions({
@@ -68,7 +97,7 @@ setOptions({
    * show horizontal addons panel as a vertical panel on the right
    * @type {Boolean}
    */
-  addonPanelInRight: true,
+  addonPanelInRight: false,
   /**
    * sorts stories
    * @type {Boolean}
@@ -83,13 +112,9 @@ setOptions({
    *   /\/|\./ - split by `/` or `.`
    * @type {Regex}
    */
-  hierarchySeparator: /\/|\./,
-
-  /**
-   * sidebar tree animations
-   * @type {Boolean}
-   */
-  sidebarAnimations: true,
+  hierarchySeparator: /\//,
+  hierarchyRootSeparator: /\|/,
+  sidebarAnimations: false,
 
   /**
    * id to select an addon panel