+++ src/components/AppFooter/AppFooter.stories.js
@@ -4,10 +4,10 @@ import { text } from "@storybook/addon-knobs/react";
 import AppFooter from "./AppFooter";
 import { withInfo } from "@storybook/addon-info";
 
-const stories = storiesOf("AppFooter", module);
+const stories = storiesOf("Components|App Footer", module);
 
 stories.add(
-  "Full Width",
+  "default",
   withInfo(
     "AppFooter for Decipher products. You can supply a `copyrightText` prop which can be an element, node, or string. This can be useful if your app supports i18n. Otherwise, the AppFooter will default to the English copyright text: 'Copyright &copy; 2018 Decipher Technology Studios. All rights reserved. Copyright &copy; 2018 Grey Matter, a Decipher Technology Studios product. All rights reserved.'"
   )(() => <AppFooter copyrightText={text("copyrightText")} />)