+++ src/components/AppHeader/components/HeaderToolbar/components/HeaderToolbarWrapper.js
@@ -1,21 +1,29 @@
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 
-import { readableColor, transparentize } from 'polished';
+import { readableColor, transparentize } from "polished";
 
 import { spacingScale } from "style/styleFunctions";
-import { FONT_SIZE_BASE, COLOR_BACKGROUND_C } from "style/styleVariables";
 
 const APP_TOOLBAR_HEIGHT = spacingScale(4);
-const APP_TOOLBAR_FONT_SIZE = FONT_SIZE_BASE;
+const APP_TOOLBAR_FONT_SIZE = props => props.theme.FONT_SIZE_BASE;
 const APP_TOOLBAR_BACKGROUND_COLOR = "transparent";
-const APP_TOOLBAR_TEXT_COLOR = readableColor(COLOR_BACKGROUND_C);
+const APP_TOOLBAR_TEXT_COLOR = props =>
+  readableColor(props.theme.COLOR_BACKGROUND_C);
+
+const backgroundImage = css`
+  background-image: linear-gradient(
+    to top,
+    ${props => transparentize(1, props.theme.COLOR_BACKGROUND_C)},
+    ${props => props.theme.COLOR_BACKGROUND_C}
+  );
+`;
 
 const HeaderToolbarWrapper = styled.nav`
+  ${backgroundImage};
   background-color: ${APP_TOOLBAR_BACKGROUND_COLOR};
   color: ${APP_TOOLBAR_TEXT_COLOR};
   flex: 0 0 ${APP_TOOLBAR_HEIGHT};
   font-size: ${APP_TOOLBAR_FONT_SIZE};
-  background-image: linear-gradient(to top, ${transparentize(1, COLOR_BACKGROUND_C)}, ${COLOR_BACKGROUND_C});
   padding: 0 ${spacingScale(2)};
   align-items: center;
   display: flex;